Split scalars tests to dedicated file
